refactor(config): move ConfigKeyAttribute to its own file
